float

+++ math.js
@@ -3,5 +3,5 @@
 exports.isStar = true;
 
 exports.sum = function (a, b) {
-    return parseInt(a) + parseInt(b);
+    return parseFloat(a) + parseFloat(b);
 };
